Use Action Env to access request body
We lazily consume the request body in the ResourceT of 'execAction',
before running the Action, and then pass it on as part of the
environment. (This is also the way Scotty does it.) Because the Action
monad itself is not a ResourceT, it would otherwise be tricky
(impossible?) to get at the request body.

Of course, we could have made Action a ResourceT or something, but I
haven't really thought that through yet..

The Env data type also gives us the advantage of "caching" query
parameters that have been converted to Text and possibly other stuff.